Validate Manager and Role

+++ ABInBev.Employees/ABInBev.Employees.Business/Services/EmployeeService.cs
@@ -16,9 +16,12 @@ namespace ABInBev.Employees.Business.Services
             _userManager = userManager;
         }
 
-        public async Task AddAsync(Employee employee, string password)
+        public async Task AddAsync(Employee employee, string password, string emailAuthenticatedUser)
         {
             await ValidatorHelper.ValidateAsync(new EmployeeValidator(_repository, null), employee);
+            var authenticatedUser = await _repository.GetByEmailAsync(emailAuthenticatedUser);
+            if (!authenticatedUser.CanUseRole(employee.Role))
+                throw new InvalidOperationException($"You are not allowed to create employees with role {employee.Role}.");
 
             var user = new IdentityUser
             {
@@ -39,13 +42,18 @@ namespace ABInBev.Employees.Business.Services
             await _repository.AddAsync(employee);
         }
 
-        public async Task UpdateAsync(Employee employee)
+        public async Task UpdateAsync(Employee employee, string emailAuthenticatedUser)
         {
             await ValidatorHelper.ValidateAsync(new EmployeeValidator(_repository, employee.Id), employee);
 
             var employeeDb = await _repository.GetByIdAsync(employee.Id);
             if (employeeDb is null)
                 throw new InvalidOperationException($"The Employee {employee.Id} was not found.");
+            var authenticatedUser = await _repository.GetByEmailAsync(emailAuthenticatedUser);
+            if (!authenticatedUser.CanUseRole(employee.Role))
+                throw new InvalidOperationException($"You are not allowed to edit employees with role {employee.Role}.");
+            if (employee.Role != employeeDb.Role && !authenticatedUser.CanUseRole(employeeDb.Role))
+                throw new InvalidOperationException($"You are not allowed to edit employees with role {employeeDb.Role}.");
 
             employeeDb.BirthDate = employee.BirthDate;
             employeeDb.DocumentNumber = employee.DocumentNumber;
@@ -58,10 +66,14 @@ namespace ABInBev.Employees.Business.Services
             await _repository.UpdateAsync(employeeDb);
         }
 
-        public async Task DeleteAsync(Guid id)
+        public async Task DeleteAsync(Guid id, string emailAuthenticatedUser)
         {
             var employee = await GetByIdAsync(id);
             if (employee is null) return;
+            
+            var authenticatedUser = await _repository.GetByEmailAsync(emailAuthenticatedUser);
+            if (!authenticatedUser.CanUseRole(employee.Role))
+                throw new InvalidOperationException($"You are not allowed to delete employees with role {employee.Role}.");
 
             await _repository.DeleteAsync(id);
 
@@ -78,6 +90,11 @@ namespace ABInBev.Employees.Business.Services
         public async Task<Employee?> GetByIdAsync(Guid id)
         {
             return await _repository.GetByIdAsync(id);
+        }
+
+        public async Task<Employee?> GetByEmailAsync(string email)
+        {
+            return await _repository.GetByEmailAsync(email);
         }
     }
 }
